Some code review fixes:
Convert the readme into asciidoc -- but still need info about icon and
category for this tls-scripts module.

Start changing the generate-trust-stores script to not download
package-kafka, but waiting on feedback first.

Download package-kafka in the docker with-secrets build, waiting on
feedback.

Use dedicated-test-key in test that creates tls certs, and in test
that generates trust stores.

Remove unneeded env-vars for download-rds-cert test, and use
require.FileExistsf instead of files.FileExists.