Require 0.12.26 everywhere except with K8S
This PR updates the `aws-service-catalog` to allow `0.12.26` or above—which includes TF 0.13—for all modules _except_ for anything K8S related, as we know `terraform-aws-eks` has an incompatibility with TF 0.13 and we are still working on a fix. This change will at least allow our non EKS users to use TF 0.13 though.