Fix RenovateBot config
While working on adding RenovateBot to the Architecture Catalog, I noticed that version updates for code in public GitHub repos—e.g., the version for Terraform and Hub—were not working. This is because ALL requests were being sent to our hacky Gruntwork Registry, which currently can only handle requests for private Gruntwork repos.

This PR does the following:

1. Update the RenovateBot config to only send `terraform-module` requests to the Registry, as piggybacking on the ability to specify a custom registry is how we get RenovateBot to send all requests for private Gruntwork repo versions (not just Terraform) to our Registry. With this update, the `github-releases` and `github-tags` requests will go straight to github.com, as they should.
1. Update a couple private repos to use matchers that will send them to our Registry.