Implement PR suggestion - Return string not a list
Finally I was able to implement the return of a single value from
dynamically generated values. It is a workaround while we cannot use
`one()` (0.15+), it consists of creating a map with a single element
(that can be empty) and using `lookup()` to get this element or `null`
if there are no elements on the map.