Introduce sensitive param/function. Require 0.15 and above.
TF 0.15 throws an error if you output a sensitive parameter without marking it as sensitive. However, the sensitive attribute was only added in 0.14, so we have to require at least that version to also support 0.15. However, as we read some data from AWS Secrets Manager, we have to mark parts of that data (e.g., port number) as NOT sensitive. That requires using the nonsenstive function, which was only added in 0.15.