Add sensitive / nonsensitive to aurora module
Fix the `aurora` module so it works with Terraform 0.15: in particular, mark the `password` input as sensitive, and mark the data we read from AWS Secrets Manager _other_ than the password (e.g., port) with `nonsensitive`. In the [Terraform 0.15 upgrade](https://github.com/gruntwork-io/terraform-aws-service-catalog/pull/678), I did the same fix for the `rds` module, but missed it for the `aurora` module, and our tests didn't complain. Are there no tests for the `aurora` module?